Calculating the basis and rotation stuff is working correctly, just need texture mapping

+++ src/engine.c
@@ -19,8 +19,8 @@ typedef struct {
   ABC_FIFO fifo;
   ForeignFunctionMap fnMap;
   ModuleMap moduleMap;
-  uint32_t width;
-  uint32_t height;
+  int32_t width;
+  int32_t height;
   mtar_t* tar;
   bool running;
   bool lockstep;
@@ -251,7 +251,7 @@ ENGINE_free(ENGINE* engine) {
 }
 
 inline internal void
-ENGINE_pset(ENGINE* engine, int64_t x, int64_t y, uint32_t c) {
+ENGINE_pset(ENGINE* engine, int32_t x, int32_t y, uint32_t c) {
   // Draw pixel at (x,y)
   int32_t width = engine->width;
   int32_t height = engine->height;
@@ -282,7 +282,7 @@ ENGINE_pset(ENGINE* engine, int64_t x, int64_t y, uint32_t c) {
 }
 
 internal void
-ENGINE_print(ENGINE* engine, char* text, int64_t x, int64_t y, uint32_t c) {
+ENGINE_print(ENGINE* engine, char* text, int32_t x, int32_t y, uint32_t c) {
   int fontWidth = 8;
   int fontHeight = 8;
   int cursor = 0;
@@ -307,18 +307,18 @@ ENGINE_print(ENGINE* engine, char* text, int64_t x, int64_t y, uint32_t c) {
 }
 
 internal void
-ENGINE_line_high(ENGINE* engine, int64_t x1, int64_t y1, int64_t x2, int64_t y2, uint32_t c) {
-  int64_t dx = x2 - x1;
-  int64_t dy = y2 - y1;
-  int64_t xi = 1;
+ENGINE_line_high(ENGINE* engine, int32_t x1, int32_t y1, int32_t x2, int32_t y2, uint32_t c) {
+  int32_t dx = x2 - x1;
+  int32_t dy = y2 - y1;
+  int32_t xi = 1;
   if (dx < 0) {
     xi = -1;
     dx = -dx;
   }
-  int64_t p = 2 * dx - dy;
+  int32_t p = 2 * dx - dy;
 
-  int64_t y = y1;
-  int64_t x = x1;
+  int32_t y = y1;
+  int32_t x = x1;
   while(y <= y2) {
     ENGINE_pset(engine, x, y, c);
     if (p > 0) {
@@ -332,18 +332,18 @@ ENGINE_line_high(ENGINE* engine, int64_t x1, int64_t y1, int64_t x2, int64_t y2,
 }
 
 internal void
-ENGINE_line_low(ENGINE* engine, int64_t x1, int64_t y1, int64_t x2, int64_t y2, uint32_t c) {
-  int64_t dx = x2 - x1;
-  int64_t dy = y2 - y1;
-  int64_t yi = 1;
+ENGINE_line_low(ENGINE* engine, int32_t x1, int32_t y1, int32_t x2, int32_t y2, uint32_t c) {
+  int32_t dx = x2 - x1;
+  int32_t dy = y2 - y1;
+  int32_t yi = 1;
   if (dy < 0) {
     yi = -1;
     dy = -dy;
   }
-  int64_t p = 2 * dy - dx;
+  int32_t p = 2 * dy - dx;
 
-  int64_t y = y1;
-  int64_t x = x1;
+  int32_t y = y1;
+  int32_t x = x1;
   while(x <= x2) {
     ENGINE_pset(engine, x, y, c);
     if (p > 0) {
@@ -357,7 +357,7 @@ ENGINE_line_low(ENGINE* engine, int64_t x1, int64_t y1, int64_t x2, int64_t y2, 
 }
 
 internal void
-ENGINE_line(ENGINE* engine, int64_t x1, int64_t y1, int64_t x2, int64_t y2, uint32_t c) {
+ENGINE_line(ENGINE* engine, int32_t x1, int32_t y1, int32_t x2, int32_t y2, uint32_t c) {
   if (llabs(y2 - y1) < llabs(x2 - x1)) {
     if (x1 > x2) {
       ENGINE_line_low(engine, x2, y2, x1, y1, c);
@@ -376,10 +376,10 @@ ENGINE_line(ENGINE* engine, int64_t x1, int64_t y1, int64_t x2, int64_t y2, uint
 }
 
 internal void
-ENGINE_circle_filled(ENGINE* engine, int64_t x0, int64_t y0, int64_t r, uint32_t c) {
-  int64_t x = 0;
-  int64_t y = r;
-  int64_t d = round(M_PI - (2*r));
+ENGINE_circle_filled(ENGINE* engine, int32_t x0, int32_t y0, int32_t r, uint32_t c) {
+  int32_t x = 0;
+  int32_t y = r;
+  int32_t d = round(M_PI - (2*r));
 
   while (x <= y) {
     ENGINE_line(engine, x0 - x, y0 + y, x0 + x, y0 + y, c);
@@ -398,10 +398,10 @@ ENGINE_circle_filled(ENGINE* engine, int64_t x0, int64_t y0, int64_t r, uint32_t
 }
 
 internal void
-ENGINE_circle(ENGINE* engine, int64_t x0, int64_t y0, int64_t r, uint32_t c) {
-  int64_t x = 0;
-  int64_t y = r;
-  int64_t d = round(M_PI - (2*r));
+ENGINE_circle(ENGINE* engine, int32_t x0, int32_t y0, int32_t r, uint32_t c) {
+  int32_t x = 0;
+  int32_t y = r;
+  int32_t d = round(M_PI - (2*r));
 
   while (x <= y) {
     ENGINE_pset(engine, x0 + x, y0 + y, c);
@@ -432,7 +432,7 @@ ellipse_getRegion(double x, double y, int32_t rx, int32_t ry) {
 }
 
 internal void
-ENGINE_ellipsefill(ENGINE* engine, int64_t x0, int64_t y0, int64_t x1, int64_t y1, uint32_t c) {
+ENGINE_ellipsefill(ENGINE* engine, int32_t x0, int32_t y0, int32_t x1, int32_t y1, uint32_t c) {
 
   // Calculate radius
   int32_t rx = (x1 - x0) / 2; // Radius on x
@@ -478,7 +478,7 @@ ENGINE_ellipsefill(ENGINE* engine, int64_t x0, int64_t y0, int64_t x1, int64_t y
 }
 
 internal void
-ENGINE_ellipse(ENGINE* engine, int64_t x0, int64_t y0, int64_t x1, int64_t y1, uint32_t c) {
+ENGINE_ellipse(ENGINE* engine, int32_t x0, int32_t y0, int32_t x1, int32_t y1, uint32_t c) {
 
   // Calcularte radius
   int32_t rx = llabs(x1 - x0) / 2; // Radius on x
@@ -531,7 +531,7 @@ ENGINE_ellipse(ENGINE* engine, int64_t x0, int64_t y0, int64_t x1, int64_t y1, u
 }
 
 internal void
-ENGINE_rect(ENGINE* engine, int64_t x, int64_t y, int64_t w, int64_t h, uint32_t c) {
+ENGINE_rect(ENGINE* engine, int32_t x, int32_t y, int32_t w, int32_t h, uint32_t c) {
   ENGINE_line(engine, x, y, x, y+h-1, c);
   ENGINE_line(engine, x, y, x+w-1, y, c);
   ENGINE_line(engine, x, y+h-1, x+w-1, y+h-1, c);
@@ -539,16 +539,16 @@ ENGINE_rect(ENGINE* engine, int64_t x, int64_t y, int64_t w, int64_t h, uint32_t
 }
 
 internal void
-ENGINE_rectfill(ENGINE* engine, int64_t x, int64_t y, int64_t w, int64_t h, uint32_t c) {
+ENGINE_rectfill(ENGINE* engine, int32_t x, int32_t y, int32_t w, int32_t h, uint32_t c) {
   int32_t width = engine->width;
   int32_t height = engine->height;
-  int64_t x1 = mid(0, x, width);
-  int64_t y1 = mid(0, y, height);
-  int64_t x2 = mid(0, x + w, width);
-  int64_t y2 = mid(0, y + h, height);
+  int32_t x1 = mid(0, x, width);
+  int32_t y1 = mid(0, y, height);
+  int32_t x2 = mid(0, x + w, width);
+  int32_t y2 = mid(0, y + h, height);
 
-  for (int64_t j = y1; j < y2; j++) {
-    for (int64_t i = x1; i < x2; i++) {
+  for (int32_t j = y1; j < y2; j++) {
+    for (int32_t i = x1; i < x2; i++) {
       ENGINE_pset(engine, i, j, c);
     }
   }
@@ -605,8 +605,8 @@ ENGINE_drawDebug(ENGINE* engine) {
   snprintf(buffer, sizeof(buffer), "%.01f fps", debug->avgFps);   // here 2 means binary
   int32_t width = engine->width;
   int32_t height = engine->height;
-  int64_t startX = width - 4*8-2;
-  int64_t startY = height - 8-2;
+  int32_t startX = width - 4*8-2;
+  int32_t startY = height - 8-2;
 
   ENGINE_rectfill(engine, startX, startY, 4*8+2, 10, 0x7F000000);
   ENGINE_print(engine, buffer, startX+1,startY+1, 0xFFFFFFFF);
@@ -626,7 +626,7 @@ ENGINE_drawDebug(ENGINE* engine) {
 }
 
 internal bool
-ENGINE_canvasResize(ENGINE* engine, uint32_t newWidth, uint32_t newHeight, uint32_t color) {
+ENGINE_canvasResize(ENGINE* engine, int32_t newWidth, int32_t newHeight, uint32_t color) {
   if (engine->width == newWidth && engine->height == newHeight) {
     return true;
   }
